requirements.txt
numpy>=1.23
Pillow>=9.2